+++ app/src/main/res/values/colors.xml
@@ -14,7 +14,7 @@
     <color name="white_two">#ffffff</color>
     <color name="pinkish_grey">#c4c4c4</color>
     <color name="greyish">#a9a9a9</color>
-    <color name="brownish_grey">6a6a6a</color>
+    <color name="brownish_grey">#6a6a6a</color>
     <color name="lemon">#fef051</color>
 
     <color name="colorBackgroundImagePickPlus">#D6D6D6</color>
